moving swagger to docs
