+++ student-registration-backend/src/docs/swagger.ts
@@ -1,12 +1,11 @@
 import swaggerUi from "swagger-ui-express";
 import { Express } from "express";
-import { openApiSpec } from "./swaggerDocs";
+import { openApiSpec } from "../docs/swaggerDocs";
 
 export default function swaggerDocs(app: Express, port: number) {
-  // Swagger page
+
   app.use("/api-docs", swaggerUi.serve, swaggerUi.setup(openApiSpec));
 
-  // Docs in JSON format
   app.get("/api-docs.json", (req, res) => {
     res.setHeader("Content-Type", "application/json");
     res.send(openApiSpec);
